AbstractBubble: handle the timeout logic
instead of exposing helper methods to start/stop the timeout, make
it implicity by setting the timeout property to oter value than 0.

this change the bubble behaviour, now the timeout only start after
the bubble is presented, rather when it's created. and the timeout
is reseted when the pointer pass over it.